Fix: Java Remove and Resource Attribute Errors
This commit fixes a three of errors associated with the Java Remove
cookbook.

To support Chef 13 features, the properties used in the
cookbook needed to be updated to include new_resource as a part of the
new Chef 13 format. Without this change, newer versions of Chef would
report a hard error of unknown resource.

The second issue resolved includes the attempt of using a log resource
within the execute resource for the uninstallation of the Java Agent.
The logging resource is not recognized within the execute resource
(rightfully so) and foodcritic flags this error with resource
attribute not recognised. This commit resolves #6.

In testing we found an additional bug where the cookbook tries to run
systemctl on pre Debian 8.0 machines. Systemctl became the standard
after Debian 8.0 Jessie. We added a conditional to ensure that we
don't attempt to call systemctl pre Debian 8.0

Ran integration tests on the following platforms:
Windows 2012r2 64 bit
Debian 7.8 64 bit
Debian 6.0.10 64 bit
Centos 7.4 64 bit

Also ran on Debian 8, even though it is not currently supported on
the Java Agent.