Support handling tar files for java linux
We provide the java agent (also known as the te_agent) in a
tar file (tgz) to our customers. This feature allows the
customer to install the agent using this cookbook by
specifying a location to either a bin file, or the compressed
tar file that is shipped.

Tested on Centos 72 and Debian 8 (only tested tar functionality
as our java agent currently does not support Debian 8).