Fix broken command options for java agent (#27)
* Fix broken command options for java agent

A couple of the optional command line arguments to the java installer
were using a deprecated way to set the properties and check the
values of the attributes passed into the recipe. This went from a
deprecation warning to a hard error in Chef 13 and above.

Tested the fix using the comprehensive test suite for linux. Tested
on Centos 7.4, 64 bit.