Java (Windows) allow installation from zip files
We provide the java agent (also known as the te_agent) in a
zip file to our customers. This feature allows the
customer to install the agent using this cookbook by
specifying a location to either an msi file, or the compressed
zip file that is shipped.

Tested on Windows 10 enterprise edition using the basic and
comprehensive test suites. Also tested that the original
behavior of the msi paths worked using these suites.