Support handling compressed files for Axon
We provide the axon agent bundle in a tar file (tgz) and a zip file
to our customers. For the linux Axon Agent, this feature allows
the customer to install the agent using this cookbook by specifying
a location to either multiple bin files for the eg driver, service
and agent, or the compressed tar file that is shipped. For the
windows Axon Agent, the user can specify either the location of
the Axon Agent MSI or the zip file that is provided.

Tested using the traditional bin file install options as well as
the tar install options on Centos 72 and Debian 8. Tested using
the Windows MSI installers directly as well as the shipped zip
files.

All kitchen tests were ran on the above platforms.

Also addressed any foodcritic / cookstyle linter errors and
warnings.